requirements.txt
absl-py==0.2.2
agate==1.6.1
agate-dbf==0.2.0
agate-excel==0.2.2
agate-sql==0.5.3
alabaster==0.7.11
alembic==0.9.10
anaconda-client==1.2.2
argcomplete==1.9.4
argh==0.26.2
asn1crypto==0.24.0
astor==0.7.1
astroid==1.6.5
astropy==3.0.3
atomicwrites==1.1.5
attrs==18.1.0
Automat==0.7.0
autopep8==1.3.5
Babel==2.6.0
backcall==0.1.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.shutil-which==3.5.1
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.6.3
bitarray==0.8.3
blaze==0.10.1
bleach==2.1.3
bokeh==0.13.0
boto==2.48.0
boto3==1.7.55
botocore==1.10.55
Bottleneck==1.2.1
browsermob-proxy==0.8.0
bz2file==0.98
Cerberus==1.2
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
chest==0.2.3
click==6.7
cloudpickle==0.5.3
clyent==1.2.1
colorama==0.3.9
colour==0.1.5
comtypes==1.1.6
configobj==5.0.6
configparser==3.5.0
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.2.2
cssselect==1.0.3
csvkit==1.0.3
cycler==0.10.0
Cython==0.28.4
cytoolz==0.9.0.1
dask==0.18.1
datashape==0.5.2
dbfread==2.0.7
decorator==4.3.0
dill==0.2.8.2
docutils==0.14
EasyProcess==0.2.3
entrypoints==0.2.3
enum34==1.1.6
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.4
Flask==1.0.2
Flask-Cors==3.0.6
funcsigs==1.0.2
future==0.16.0
futures==3.1.1
gast==0.2.0
gensim==3.5.0
geojson==2.4.0
gevent==1.3.4
gmaps==0.8.0
googlemaps==3.0.2
greenlet==0.4.13
grin==1.2.1
grpcio==1.13.0
h5py==2.8.0
HeapDict==1.0.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.7
imagesize==1.0.0
incremental==17.5.0
ipaddress==1.0.22
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isodate==0.6.0
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.1
Jinja2==2.10
jmespath==0.9.3
joblib==0.12.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.32.1
jupyterlab-launcher==0.10.0
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
keyring==13.2.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
leather==0.3.3
llvmlite==0.24.0
locket==0.2.0
lxml==4.2.5
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.3
mobilenium==0.0.4
more-itertools==4.2.0
mpmath==1.0.0
multipledispatch==0.5.0
nbconvert==5.3.1
nbformat==4.4.0
nbpresent==3.0.0
networkx==2.1
nltk==3.3
nose==1.3.7
notebook==5.5.0
numba==0.39.0
numexpr==2.6.5
numpy==1.14.5
numpydoc==0.8.0
odo==0.5.0
openpyxl==2.5.4
packaging==17.1
pandas==0.23.3
pandocfilters==1.4.2
parsedatetime==2.4
parsel==1.5.0
parso==0.3.1
partd==0.3.8
path.py==11.0.1
pathlib2==2.3.2
pathtools==0.1.2
patsy==0.5.0
pep8==1.7.1
pickleshare==0.7.4
Pillow==5.2.0
pip-chill==0.1.8
pkginfo==1.4.2
pluggy==0.6.0
ply==3.11
prompt-toolkit==1.0.15
protobuf==3.6.0
psutil==5.4.6
psycopg2==2.7.5
py==1.5.4
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycurl==7.43.0.2
PyDispatcher==2.0.5
PyExcelerate==0.7.3
pyflakes==2.0.0
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.6.4
pylint==1.9.2
pyOpenSSL==18.0.0
pyparsing==2.2.0
PyQt5==5.9.2
pyreadline==2.1
pytest==3.6.3
python-dateutil==2.7.3
python-editor==1.0.3
python-slugify==1.2.5
pytimeparse==1.1.8
pytz==2018.5
PyVirtualDisplay==0.2.1
PyWavelets==0.5.2
pywin32==223
pywin32-ctypes==0.1.2
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.2
queuelib==1.5.0
requests==2.19.1
rope==0.10.7
ruamel.yaml==0.15.42
s3transfer==0.1.13
scikit-image==0.14.0
scikit-learn==0.19.1
scikit-surprise==1.0.6
scipy==1.1.0
Scrapy==1.5.1
seaborn==0.8.1
selenium==3.14.0
Send2Trash==1.5.0
service-identity==17.0.0
sh==1.12.14
simplegeneric==0.8.1
singledispatch==3.4.0.3
sip==4.19.8
six==1.11.0
sklearn==0.0
smart-open==1.6.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
Sphinx==1.7.5
sphinxcontrib-websupport==1.1.0
splinter==0.9.0
spyder==3.3.0
spyder-kernels==0.2.4
SQLAlchemy==1.2.11
statsmodels==0.9.0
subprocess32==3.5.2
surprise==0.1
sympy==1.2
tables==3.4.4
tabulate==0.8.2
tensorboard==1.9.0
tensorflow-gpu==1.9.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
Theano==1.0.2
toolz==0.9.0
tornado==5.1
traitlets==4.3.2
Twisted==18.7.0
ujson==1.35
unicodecsv==0.14.1
Unidecode==1.0.22
urllib3==1.23
w3lib==1.19.0
watchdog==0.8.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
win-unicode-console==0.5
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.5
xlwings==0.11.8
xlwt==1.3.0
yapf==0.22.0
zope.interface==4.5.0
